Extract FacilityListItemBuilder.Builder to its own file
