+++ app/src/main/java/org/simple/clinic/facility/change/FacilityListItem.kt
@@ -23,27 +23,4 @@ data class FacilityListItem(
     override fun areItemsTheSame(oldItem: FacilityListItem, newItem: FacilityListItem) = oldItem.facility.uuid == newItem.facility.uuid
     override fun areContentsTheSame(oldItem: FacilityListItem, newItem: FacilityListItem) = oldItem == newItem
   }
-
-  object Builder {
-    fun build(facility: Facility, searchQuery: String): FacilityListItem {
-      val canHighlight = searchQuery.isNotBlank() && facility.name.contains(searchQuery, ignoreCase = true)
-
-      val highlightedName = if (canHighlight) {
-        Name.Highlighted(
-            name = facility.name,
-            highlightStart = facility.name.indexOf(searchQuery, ignoreCase = true),
-            highlightEnd = facility.name.indexOf(searchQuery, ignoreCase = true) + searchQuery.length)
-      } else {
-        Name.Plain(facility.name)
-      }
-
-      val fullAddress = if (facility.streetAddress.isNullOrBlank()) {
-        Address.WithoutStreet(district = facility.district, state = facility.state)
-      } else {
-        Address.WithStreet(street = facility.streetAddress!!, district = facility.district, state = facility.state)
-      }
-
-      return FacilityListItem(facility = facility, name = highlightedName, address = fullAddress)
-    }
-  }
 }
